requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.35.2
torch>=2.2.0
huggingface-hub==0.19.4
mcp>=1.0.0
pydantic>=2.11.0
python-dotenv==1.0.0

